+++ docs/CHANGELOG.md
@@ -24,6 +24,11 @@
 
 2. 增加组件库 [ant](https://ant.design/docs/react/introduce-cn)
 
-3. css处理器
+3. css处理器(使用postcss替换sass)
+* [postcss插件列表](https://github.com/postcss/postcss/blob/main/docs/plugins.md)
+* [PostCSS](https://github.com/postcss/postcss/blob/main/docs/README-cn.md) css转换工具
+* [postcss-preset-env](https://github.com/csstools/postcss-plugins/tree/main/plugin-packs/postcss-preset-env)转化css的新特性到目标浏览器（包含autoprefixer）
+* [postcss-nested](https://github.com/postcss/postcss-nested)像sass一样的切套规则
+* [postcss-import](https://github.com/postcss/postcss-import)css中使用@import